Lib/imghdr.py
"""Recognize image file formats based on their first few bytes."""

__all__ = ["what"]

#-------------------------#
# Recognize image headers #
#-------------------------#

def what(file, h=Nic):
    f = Nic
    spróbuj:
        jeżeli h jest Nic:
            jeżeli isinstance(file, str):
                f = open(file, 'rb')
                h = f.read(32)
            inaczej:
                location = file.tell()
                h = file.read(32)
                file.seek(location)
        dla tf w tests:
            res = tf(h, f)
            jeżeli res:
                zwróć res
    w_końcu:
        jeżeli f: f.close()
    zwróć Nic


#---------------------------------#
# Subroutines per image file type #
#---------------------------------#

tests = []

def test_jpeg(h, f):
    """JPEG data w JFIF albo Exjeżeli format"""
    jeżeli h[6:10] w (b'JFIF', b'Exif'):
        zwróć 'jpeg'

tests.append(test_jpeg)

def test_png(h, f):
    jeżeli h.startswith(b'\211PNG\r\n\032\n'):
        zwróć 'png'

tests.append(test_png)

def test_gif(h, f):
    """GIF ('87 oraz '89 variants)"""
    jeżeli h[:6] w (b'GIF87a', b'GIF89a'):
        zwróć 'gif'

tests.append(test_gif)

def test_tiff(h, f):
    """TIFF (can be w Motorola albo Intel byte order)"""
    jeżeli h[:2] w (b'MM', b'II'):
        zwróć 'tiff'

tests.append(test_tiff)

def test_rgb(h, f):
    """SGI image library"""
    jeżeli h.startswith(b'\001\332'):
        zwróć 'rgb'

tests.append(test_rgb)

def test_pbm(h, f):
    """PBM (portable bitmap)"""
    jeżeli len(h) >= 3 oraz \
        h[0] == ord(b'P') oraz h[1] w b'14' oraz h[2] w b' \t\n\r':
        zwróć 'pbm'

tests.append(test_pbm)

def test_pgm(h, f):
    """PGM (portable graymap)"""
    jeżeli len(h) >= 3 oraz \
        h[0] == ord(b'P') oraz h[1] w b'25' oraz h[2] w b' \t\n\r':
        zwróć 'pgm'

tests.append(test_pgm)

def test_ppm(h, f):
    """PPM (portable pixmap)"""
    jeżeli len(h) >= 3 oraz \
        h[0] == ord(b'P') oraz h[1] w b'36' oraz h[2] w b' \t\n\r':
        zwróć 'ppm'

tests.append(test_ppm)

def test_rast(h, f):
    """Sun raster file"""
    jeżeli h.startswith(b'\x59\xA6\x6A\x95'):
        zwróć 'rast'

tests.append(test_rast)

def test_xbm(h, f):
    """X bitmap (X10 albo X11)"""
    jeżeli h.startswith(b'#define '):
        zwróć 'xbm'

tests.append(test_xbm)

def test_bmp(h, f):
    jeżeli h.startswith(b'BM'):
        zwróć 'bmp'

tests.append(test_bmp)

def test_webp(h, f):
    jeżeli h.startswith(b'RIFF') oraz h[8:12] == b'WEBP':
        zwróć 'webp'

tests.append(test_webp)

def test_exr(h, f):
    jeżeli h.startswith(b'\x76\x2f\x31\x01'):
        zwróć 'exr'

tests.append(test_exr)

#--------------------#
# Small test program #
#--------------------#

def test():
    zaimportuj sys
    recursive = 0
    jeżeli sys.argv[1:] oraz sys.argv[1] == '-r':
        usuń sys.argv[1:2]
        recursive = 1
    spróbuj:
        jeżeli sys.argv[1:]:
            testall(sys.argv[1:], recursive, 1)
        inaczej:
            testall(['.'], recursive, 1)
    wyjąwszy KeyboardInterrupt:
        sys.stderr.write('\n[Interrupted]\n')
        sys.exit(1)

def testall(list, recursive, toplevel):
    zaimportuj sys
    zaimportuj os
    dla filename w list:
        jeżeli os.path.isdir(filename):
            print(filename + '/:', end=' ')
            jeżeli recursive albo toplevel:
                print('recursing down:')
                zaimportuj glob
                names = glob.glob(os.path.join(filename, '*'))
                testall(names, recursive, 0)
            inaczej:
                print('*** directory (use -r) ***')
        inaczej:
            print(filename + ':', end=' ')
            sys.stdout.flush()
            spróbuj:
                print(what(filename))
            wyjąwszy OSError:
                print('*** nie found ***')

jeżeli __name__ == '__main__':
    test()
